requirements.txt
Flask==2.3.3
transformers==4.41.1
torch==2.2.0
PyPDF2==3.0.1
python-docx==0.8.11
reportlab